requirements.txt
pandas==1.3.1
scipy==1.6.2
streamlit==1.12.2
altair<5
nbformat
plotly-express